fix(IRC): rename class to reference WebsocketKit
